Modernize deployment (#66)
Update deployment techniques to use containers and kubernetes, instead of a hybrid-container approach and docker-compose.

Tilt has been added now to allow for live testing in a local Kubernetes cluster. Start the app on a kubernetes cluster with `tilt up` and then access it at `localhost:8080`.

These changes should also make it easier to use the new Github Actions CI for automatic testing and deployment, instead of Travis CI. Since most complexity has been moved to the Containerfile and Kubernetes, we no longer need to depend on the CI environment as much.